notebooks/requirements.txt
gradio
torch
transformers
pandas
numpy
yt-dlp
webvtt-py
langchain
pinecone-client
python-dotenv
datasets
nltk
openai-whisper
gTTS
evaluate
tiktoken
langchain-openai
langchain-pinecone
langchain-community
langchain==0.1.20
openai




